server/requirements.txt
advanced_alchemy==0.8.4
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
ecdsa==0.19.0
EditorConfig==0.12.4
email_validator==2.1.1
Faker==25.1.0
fast-query-parsers==1.0.3
greenlet==3.0.3
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsbeautifier==1.15.1
litestar==2.8.3
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
polyfactory==2.15.0
psycopg==3.2.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.31.0
rich==13.7.1
rich-click==1.8.1
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
